Revert "Merge branch 'puppetlabs/master'"
This reverts commit 90e4874e86f6df64a3b759ee673e8267768e70d3, reversing
changes made to 5e1649a73662dc519796598d8d3d10485b00beca.

+++ lib/puppet/provider/route53_zone/v2.rb
@@ -42,8 +42,10 @@ Puppet::Type.type(:route53_zone).provide(:v2, :parent => PuppetX::Puppetlabs::Aw
   def destroy
     Puppet.info("Deleting zone #{name}")
     zones = route53_client.list_hosted_zones.data.hosted_zones.select { |zone| zone.name == name }
-    zones.each do |zone|
-      route53_client.delete_hosted_zone(id: zone.id)
+    if zones
+      zones.each do |zone|
+        route53_client.delete_hosted_zone(id: zone.id)
+      end
     end
     @property_hash[:ensure] = :absent
   end
